Add files via upload
One Fish, Two Fish, Red Fish, Blue Fish......

A fun exercise which helped me understand defining functions and why planning your program and making it modular is a good idea. By defining so much within the drawFish() function, we are able to simply feed the function a few parameters in order to draw a new fish.